+++ src/main/kotlin/blackjack/model/Dealer.kt
@@ -1,5 +1,8 @@
 package blackjack.model
 
+import blackjack.model.card.CardHand
+import blackjack.model.card.CardHandState
+
 data class Dealer(override val cardHand: CardHand) : Role(name = PlayerName(DEALER), cardHand) {
     override fun getState(): CardHandState {
         val sum = cardHand.sum()
