+++ src/main/kotlin/blackjack/model/card/Card.kt
@@ -1,3 +1,3 @@
-package blackjack.model
+package blackjack.model.card
 
 data class Card(val shape: CardShape, val number: CardNumber)
