+++ src/main/kotlin/blackjack/model/card/CardHandState.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 enum class CardHandState {
     BLACKJACK,
