+++ src/main/kotlin/blackjack/model/card/CardShape.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 enum class CardShape(
     val shape: String,
