+++ src/test/kotlin/blackjack/model/DealerTest.kt
@@ -1,5 +1,10 @@
 package blackjack.model
 
+import blackjack.model.card.Card
+import blackjack.model.card.CardHand
+import blackjack.model.card.CardHandState
+import blackjack.model.card.CardNumber
+import blackjack.model.card.CardShape
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
